+++ contracts/interfaces/IUniswapV3Factory.sol
@@ -3,8 +3,8 @@ pragma solidity >=0.5.0;
 
 interface IUniswapV3Factory {
     event OwnerChanged(address indexed oldOwner, address indexed newOwner);
-    event PairCreated(address indexed token0, address indexed token1, uint16 indexed fee, address pair, uint256);
-    event FeeOptionEnabled(uint16 indexed fee);
+    event PairCreated(address indexed token0, address indexed token1, uint24 indexed fee, address pair, uint256);
+    event FeeOptionEnabled(uint24 indexed fee);
 
     function owner() external view returns (address);
 
@@ -12,25 +12,25 @@ interface IUniswapV3Factory {
 
     function allPairsLength() external view returns (uint256);
 
-    function allEnabledFeeOptions(uint256) external view returns (uint16);
+    function allEnabledFeeOptions(uint256) external view returns (uint24);
 
     function allEnabledFeeOptionsLength() external view returns (uint256);
 
     function getPair(
         address tokenA,
         address tokenB,
-        uint16 fee
+        uint24 fee
     ) external view returns (address pair);
 
-    function isFeeOptionEnabled(uint16 fee) external view returns (bool);
+    function isFeeOptionEnabled(uint24 fee) external view returns (bool);
 
     function createPair(
         address tokenA,
         address tokenB,
-        uint16 fee
+        uint24 fee
     ) external returns (address pair);
 
     function setOwner(address) external;
 
-    function enableFeeOption(uint16 fee) external;
+    function enableFeeOption(uint24 fee) external;
 }
